+++ .dim/teardown.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env sh
+set -eu
+
+docker compose --project-name "dim-${DIM_WORKSPACE_NAME}" \
+  --file .dim/docker-compose.yml down --volumes --remove-orphans