+++ examples/project/repos/root/.dim/setup.sh
@@ -1,0 +1,13 @@
+#!/usr/bin/env sh
+set -eu
+
+git_name="$(dim-host-input builtin.git-author name)"
+git_email="$(dim-host-input builtin.git-author email)"
+
+export GIT_AUTHOR_NAME="$git_name"
+export GIT_AUTHOR_EMAIL="$git_email"
+export GIT_COMMITTER_NAME="$git_name"
+export GIT_COMMITTER_EMAIL="$git_email"
+
+docker compose --project-name "dim-${DIM_WORKSPACE_NAME}" \
+  --file .dim/docker-compose.yml up --detach --build agent